% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{add_dataset_if_not_exists}
\alias{add_dataset_if_not_exists}
\title{Add dataset if not exists}
\usage{
add_dataset_if_not_exists(project_hid, filename, title,
  prediction_only = FALSE)
}
\arguments{
\item{project_hid}{character with project identifier}

\item{filename}{character with filename containing data}

\item{title}{title of dataset}

\item{prediction_only}{boolean determining if data is used only for prediction}
}
\value{
parsed dataset details
}
\description{
Checks parameters before adding new dataset and verifies
if it doesn't exists already.
}
