% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{add_experiment_if_not_exists}
\alias{add_experiment_if_not_exists}
\title{Add experiment if not exists}
\usage{
add_experiment_if_not_exists(project_hid, train_dataset, valid_dataset,
  experiment_title, project_task, validation_kfolds, validation_shuffle,
  validation_stratify, validation_train_split, algorithms, metric, tuning_mode,
  time_constraint, create_ensemble)
}
\arguments{
\item{project_hid}{character with project identifier}

\item{train_dataset}{character with path to training dataset}

\item{valid_dataset}{character with path to validation dataset}

\item{experiment_title}{character with experiment title}

\item{project_task}{character with project task}

\item{validation_kfolds}{number of folds to be used in validation}

\item{validation_shuffle}{boolean which specify if shuffle samples before training}

\item{validation_stratify}{boolean which decides whether samples will be
divided into folds with the same class distribution}

\item{validation_train_split}{ratio how to split training dataset into train and validation}

\item{algorithms}{list of algorithms to use}

\item{metric}{charcater with metric}

\item{tuning_mode}{tuning mode}

\item{time_constraint}{numeric with time limit to calculate algorithm}

\item{create_ensemble}{whether or not to create ensemble}
}
\value{
experiment details structure
}
\description{
Check if experiment exists, verifies parameters, creates data
to create_experiment function and finally starts creation of
MLJAR experiment.
}
