% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{add_new_dataset}
\alias{add_new_dataset}
\title{Adds new dataset}
\usage{
add_new_dataset(project_hid, filename, title, prediction_only = FALSE)
}
\arguments{
\item{project_hid}{character with project identifier}

\item{filename}{character with filename containing data}

\item{title}{title of dataset}

\item{prediction_only}{boolean determining if data is used only for prediction}
}
\value{
parsed by toJSON dataset details
}
\description{
Adds new dataset
}
