% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{create_experiment}
\alias{create_experiment}
\title{Creates experiment from given parameters}
\usage{
create_experiment(data)
}
\arguments{
\item{data}{list of experiment parameters}
}
\value{
experiment details parsed by fromJSON
}
\description{
Creates experiment from given parameters
}
