% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{create_project}
\alias{create_project}
\title{Creates a new project}
\usage{
create_project(title, task, description = "")
}
\arguments{
\item{title}{character with project title}

\item{task}{character with project task}

\item{description}{optional description}
}
\value{
project details structure
}
\description{
Creates a new project
}
