% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_all_models}
\alias{get_all_models}
\title{Gives data.frame with basic data of all models}
\usage{
get_all_models(project_title, exp_title)
}
\arguments{
\item{project_title}{character with project title}

\item{exp_title}{character with experiment title}
}
\value{
data.frame with model's "hid", "model_type", "metric_value",
"metric_type"
}
\description{
You can later get some specific model by calling
e.g. `get_result(model_hid)'.
}
