% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_dataset}
\alias{get_dataset}
\title{Gets dataset}
\usage{
get_dataset(dataset_hid)
}
\arguments{
\item{dataset_hid}{character with dataset identifier}
}
\value{
structure with parsed dataset and response
}
\description{
Gets dataset
}
