% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_datasets}
\alias{get_datasets}
\title{Gets list of available datasets}
\usage{
get_datasets(project_hid)
}
\arguments{
\item{project_hid}{character with project identifier}
}
\value{
structure with parsed datasets and response
}
\description{
Gets list of available datasets
}
