% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{get_experiment}
\alias{get_experiment}
\title{Gets experiment details}
\usage{
get_experiment(experiment_hid)
}
\arguments{
\item{experiment_hid}{character with experiment identifier}
}
\value{
structure with parsed experiment and http response
}
\description{
Gets experiment details
}
