% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{get_experiments}
\alias{get_experiments}
\title{Gets list of available experiments for given project}
\usage{
get_experiments(project_hid)
}
\arguments{
\item{project_hid}{character with project identifier}
}
\value{
structure with parsed experiments and http response
}
\description{
Gets list of available experiments for given project
}
