% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{get_model}
\alias{get_model}
\title{Get model}
\usage{
get_model(project_title, exp_title, model_hid)
}
\arguments{
\item{project_title}{character with project title}

\item{exp_title}{character with experiment title}

\item{model_hid}{character with experiment identifier}
}
\value{
structure with model parameters
}
\description{
Gets model only if experiment finished and project with such
a title and having such an experiment exists.
}
