% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{get_prediction}
\alias{get_prediction}
\title{Gets MLJAR predictions}
\usage{
get_prediction(project_hid, dataset_hid, result_hid)
}
\arguments{
\item{project_hid}{character with project identifier}

\item{dataset_hid}{character with dataset identifier}

\item{result_hid}{character with result identifier}
}
\value{
structure with parsed prediction and http response
}
\description{
Gets MLJAR predictions
}
