% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{get_project}
\alias{get_project}
\title{Get project}
\usage{
get_project(hid)
}
\arguments{
\item{hid}{character with project unique identifier}
}
\value{
structure with parsed project and http response
}
\description{
Get data from a project of specified hid
}
