% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{get_results}
\alias{get_results}
\title{#' Get results of MLJAR training}
\usage{
get_results(project_hid, experiment_hid)
}
\arguments{
\item{project_hid}{character with project identifier}

\item{experiment_hid}{character with experiment identifier}
}
\value{
structure with parsed results and http response
}
\description{
#' Get results of MLJAR training
}
