% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mljar_predict}
\alias{mljar_predict}
\title{MLJAR PREDICT}
\usage{
mljar_predict(model, x_pred, project_title)
}
\arguments{
\item{model}{model or MLJAR result structure}

\item{x_pred}{data.frame/matrix data to predict}

\item{project_title}{character with project title}
}
\value{
data.frame with preditction
}
\description{
Makes prediction basing on trained model.
}
