% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_download.R
\name{prediction_download}
\alias{prediction_download}
\title{Function to get predictions from MLJAR.}
\usage{
prediction_download(prediction_hid)
}
\arguments{
\item{prediction_hid}{prediction identifier}
}
\value{
data.frame with prediction
}
\description{
Function to get predictions from MLJAR.
}
