% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{print_all_projects}
\alias{print_all_projects}
\title{Print all projects}
\usage{
print_all_projects()
}
\value{
data.frame with projects
}
\description{
Gives data.frame with basic information about existing projects
}
