% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictjob.R
\name{submit_predict_job}
\alias{submit_predict_job}
\title{Submits dataset for MLJAR prediction}
\usage{
submit_predict_job(project_hid, dataset_hid, result_hid)
}
\arguments{
\item{project_hid}{character with project identifier}

\item{dataset_hid}{character with dataset identifier}

\item{result_hid}{character with result identifier}
}
\description{
Submits dataset for MLJAR prediction
}
