% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataupload.R
\name{upload_file}
\alias{upload_file}
\title{Uploads file into MLJAR}
\usage{
upload_file(project_hid, filepath)
}
\arguments{
\item{project_hid}{character with project identifier}

\item{filepath}{character with path to file}
}
\value{
character with destination path
}
\description{
It uploads file into MLJAR and returns destination path.
}
